pyproject.toml
[project]
name = "tool-eval"
version = "0.1.0"
description = "Eval harness for LLM tool calling capabilities"
requires-python = ">=3.11"
dependencies = [
    "instructor>=1.7.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ipython>=8.0.0",
]

[project.scripts]
tool-eval = "tool_eval.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tool_eval"]
